backend/requirements.txt
# ─── Web Framework ────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
websockets>=13.0
httpx>=0.27.0

# ─── Database ─────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# ─── Cache ────────────────────────────────────────────────────────────────────
redis[hiredis]>=5.1.0

# ─── Auth & Security ──────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=43.0.0

# ─── LLM / AI ────────────────────────────────────────────────────────────────
anthropic>=0.40.0
openai>=1.54.0
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0

# ─── RAG / Vector DB ─────────────────────────────────────────────────────────
qdrant-client>=1.12.0
rank-bm25>=0.2.2
tiktoken>=0.7.0

# ─── Voice AI ─────────────────────────────────────────────────────────────────
pydub>=0.25.1

# ─── Validation ───────────────────────────────────────────────────────────────
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# ─── Observability ───────────────────────────────────────────────────────────
prometheus-client>=0.21.0
structlog>=24.4.0

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tenacity>=9.0.0
aiofiles>=24.1.0
anyio>=4.6.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
